account setup begun

+++ src/PlayGen.SGA.DataController.UnitTests/TestDbController.cs
@@ -8,25 +8,33 @@ namespace PlayGen.SGA.DataController.UnitTests
 {
     public abstract class TestDbController
     {
-        protected readonly string _nameOrConnectionString;
         public const string DbName = "sgadatacontrollerunittests";
+
+        protected readonly string _nameOrConnectionString;
+
+        private static bool _deletedDatabase = false;
 
         public TestDbController()
         {
             _nameOrConnectionString = "Server=localhost;" +
                                       "Port=3306;" +
-                                      "Database=" + DbName + ";" +
+                                      $"Database={DbName};" +
                                       "Uid=root;" +
                                       "Pwd=;" +
                                       "Convert Zero Datetime=true;" +
                                       "Allow Zero Datetime=true";
 
-            using (var context = new SGAContext(_nameOrConnectionString))
+            if (!_deletedDatabase)
             {
-                if (context.Database.Connection.Database == DbName)
+                using (var context = new SGAContext(_nameOrConnectionString))
                 {
-                    context.Database.Delete();
+                    if (context.Database.Connection.Database == DbName)
+                    {
+                        context.Database.Delete();
+                    }
                 }
+
+                _deletedDatabase = true;
             }
         }
     }
